Rename PreLoadRepository to PreLoadRecord
